api refactoring - moved api folder out of demos folder and into top level of source - changed all mentions of it to the new path - deleted some unused things from earlier
