--- source/card/api_key.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line no-unused-vars
-const API_KEY = "d74e3103-e71a-4db1-a29b-ac1955aa81cd"
