--- source/demos/api-search/.gitignore
@@ -1,1 +1,0 @@
-api_key.js
